🔨 integrate DiscreteBars into DiscreteBarChart
